+++ src/qz/printer/action/html/WebApp.java
@@ -364,224 +364,8 @@ public class WebApp extends Application {
         }
 
         Platform.runLater(() -> {
-            final Stage previewStage = new Stage(stage.getStyle());
-            final WebView webPreview = new WebView();
-
-            if (model.isPlainText()) {
-                webPreview.getEngine().loadContent(model.getSource(), "text/html");
-            } else {
-                webPreview.getEngine().load(model.getSource());
-            }
-
-            int thickness = 20;
-            double dpi = 72;
-
-            Canvas leftRuler = new Canvas(thickness, 400);
-            Canvas topRuler = new Canvas(400, thickness);
-
-            drawLeftRuler(leftRuler, thickness, leftRuler.getHeight(), dpi);
-            drawTopRuler(topRuler, thickness, topRuler.getWidth(), dpi);
-
-            StackPane webContainer = new StackPane(webPreview);
-            webPreview.prefWidthProperty().bind(webContainer.widthProperty());
-            webPreview.prefHeightProperty().bind(webContainer.heightProperty());
-
-            final BorderPane borderPane = new BorderPane();
-            borderPane.setTop(topRuler);
-            borderPane.setLeft(leftRuler);
-            borderPane.setCenter(webContainer);
-
-            final Label info = new Label("Resize me!");
-            final TextField widthField = new TextField();
-            final TextField heightField = new TextField();
-            final Label unit = new Label("in");
-
-            widthField.setPrefColumnCount(3);
-            heightField.setPrefColumnCount(3);
-
-            HBox spring = new HBox();
-            HBox.setHgrow(spring, Priority.ALWAYS);
-
-            ImageView cancelIcon = new ImageView(SwingFXUtils.toFXImage(IconCache.getInstance().getImage(IconCache.Icon.CANCEL_ICON), null));
-            ImageView doneIcon = new ImageView(SwingFXUtils.toFXImage(IconCache.getInstance().getImage(IconCache.Icon.ALLOW_ICON), null));
-            final Button cancel = new Button("Cancel", cancelIcon);
-            final Button done = new Button("Done", doneIcon);
-            done.setAlignment(Pos.CENTER_RIGHT);
-
-            final ToolBar toolBar = new ToolBar(
-                    info,
-                    widthField,
-                    heightField,
-                    unit,
-                    spring, //for spacing
-                    cancel,
-                    done
-            );
-
-            final BorderPane toolbarPane = new BorderPane();
-            toolbarPane.setTop(toolBar);
-            toolbarPane.setCenter(borderPane);
-
-            //final Label info = new Label("hello world");
-            ////todo big no on this color. possibly use a generic color to avoid white label constant?
-            //info.setBackground(Background.fill(Color.web("#80FF80", .5)));
-            //info.setPadding(new Insets(5));
-            //StackPane guiLayer = new StackPane(toolbarPane, info);
-
-            //StackPane.setAlignment(info, Pos.BOTTOM_RIGHT);
-            //StackPane.setMargin(info, new Insets(20));
-
-            Scene scene = new Scene(toolbarPane);
-            previewStage.setTitle("HTML Preview - " + settings.getJobName());
-            previewStage.setScene(scene);
-            previewStage.sizeToScene();
-            previewStage.show();
-
-            Platform.runLater(() -> {
-                previewStage.toFront();
-                info.requestFocus();
-            });
-
-            widthField.focusedProperty().addListener((observable, oldValue, newValue) -> {
-                if (newValue) return; //we just gained focus, we don't need to parse any input yet
-
-                double newWidth = parseInput(widthField.getText());
-                if (newWidth > 0) {
-                    log.warn("changing width");
-                    // The stage size and scene size are not the same. I think this is due to the window border. I could not find a more direct approach.
-                    double fudgeFactor = previewStage.getWidth() - scene.getWidth();
-                    previewStage.setWidth(newWidth * dpi + thickness + fudgeFactor);
-                } else {
-                    DecimalFormat nf = new DecimalFormat("#.##");
-                    widthField.setText(nf.format((scene.getWidth() - thickness) / dpi));
-                }
-            });
-
-            heightField.focusedProperty().addListener((observable, oldValue, newValue) -> {
-                if (newValue) return; //we just gained focus, we don't need to parse any input yet
-
-                double newHeight = parseInput(heightField.getText());
-                if (newHeight > 0) {
-                    log.warn("changing width");
-                    // The stage size and scene size are not the same. I think this is due to the window border. I could not find a more direct approach.
-                    double fudgeFactor = previewStage.getHeight() - scene.getHeight();
-                    previewStage.setHeight(newHeight * dpi + thickness + toolBar.getHeight() + fudgeFactor);
-                } else {
-                    DecimalFormat nf = new DecimalFormat("#.##");
-                    heightField.setText(nf.format((scene.getHeight() - thickness - toolBar.getHeight()) / dpi));
-                }
-            });
-
-            widthField.setOnKeyPressed(keyEvent -> {
-                if (keyEvent.getCode() == KeyCode.ENTER) {
-                    info.requestFocus();
-                }
-                if (keyEvent.getCode() == KeyCode.ESCAPE) {
-                    DecimalFormat nf = new DecimalFormat("#.##");
-                    widthField.setText(nf.format((scene.getWidth() - thickness) / dpi));
-                    info.requestFocus();
-                }
-            });
-
-            heightField.setOnKeyPressed(keyEvent -> {
-                if (keyEvent.getCode() == KeyCode.ENTER) {
-                    info.requestFocus();
-                }
-                if (keyEvent.getCode() == KeyCode.ESCAPE) {
-                    DecimalFormat nf = new DecimalFormat("#.##");
-                    heightField.setText(nf.format((scene.getHeight() - thickness - toolBar.getHeight()) / dpi));
-                    info.requestFocus();
-                }
-            });
-
-            scene.widthProperty().addListener((obs, oldVal, newVal) -> {
-                double newWidth = (double)newVal;
-
-                DecimalFormat nf = new DecimalFormat("#.##");
-                widthField.setText(nf.format((newWidth - thickness) / dpi));
-                heightField.setText(nf.format((scene.getHeight() - thickness - toolBar.getHeight()) / dpi));
-                info.setText("Current: ");
-
-                topRuler.setWidth(newWidth);
-                drawTopRuler(topRuler, thickness, newWidth, dpi);
-            });
-
-            scene.heightProperty().addListener((obs, oldVal, newVal) -> {
-                double newHeight = scene.getHeight() - thickness; //the top ruler cuts this ruler off
-
-                DecimalFormat nf = new DecimalFormat("#.##");
-                info.setText("Current: ");
-                widthField.setText(nf.format((scene.getWidth() - thickness) / dpi));
-                heightField.setText(nf.format((newHeight - toolBar.getHeight()) / dpi));
-
-                leftRuler.setHeight(newHeight);
-                drawLeftRuler(leftRuler, thickness, newHeight, dpi);
-            });
-
-            scene.setOnKeyPressed(event -> {
-                System.out.println("Key pressed: " + event.getCode());
-            });
+            new PreviewHTML(model, settings, new Stage(stage.getStyle()));
         });
-    }
-
-    private static double parseInput(String value) {
-        try {
-            return Double.parseDouble(value);
-        } catch (NumberFormatException e) {
-            return -1;
-        }
-    }
-
-    private static void drawLeftRuler(Canvas canvas, int thickness, double height, double dpi) {
-        GraphicsContext gc = canvas.getGraphicsContext2D();
-        gc.clearRect(0, 0, thickness, height);
-        gc.setStroke(Color.BLACK);
-        gc.setLineWidth(1);
-        Font font = Font.font("Arial", FontWeight.LIGHT, 8);
-        gc.setFont(font);
-
-        double spacing = dpi / 8;
-        for (int i = 1; i * spacing < height; i++) {
-            double lineHeight = thickness * 0.2;
-            if (i % 4 == 0) lineHeight += thickness * 0.3;
-            if (i % 8 == 0) {
-                Text helper = new Text(String.valueOf(i / 8));
-                helper.setFont(font);
-                double textWidth = Math.ceil(helper.getLayoutBounds().getWidth());
-
-                gc.save();
-                gc.translate(0, i * spacing);
-                gc.rotate(-90);
-                gc.strokeText(String.valueOf(i / 8), -textWidth / 2, thickness - 2);
-                gc.restore();
-            }
-            gc.strokeLine(0, i * spacing, lineHeight, i * spacing);
-        }
-    }
-
-    private static void drawTopRuler(Canvas canvas, int thickness, double width, double dpi) {
-        GraphicsContext gc = canvas.getGraphicsContext2D();
-        gc.clearRect(0, 0, width, thickness);
-        gc.setStroke(Color.BLACK);
-        gc.setFill(Color.GRAY);
-        gc.fillRect(0,0, thickness, thickness);
-        gc.setLineWidth(1);
-        Font font = Font.font("Arial", FontWeight.LIGHT, 8);
-        gc.setFont(font);
-
-        double spacing = dpi / 8;
-        for (int i = 1; i * spacing < width; i++) {
-            double lineHeight = thickness * 0.2;
-            if (i % 4 == 0) lineHeight += thickness * 0.3;
-            if (i % 8 == 0) {
-                Text helper = new Text(String.valueOf(i / 8));
-                helper.setFont(font);
-                double textWidth = Math.ceil(helper.getLayoutBounds().getWidth());
-                gc.strokeText(String.valueOf(i / 8),
-                              thickness + i * spacing - (textWidth / 2), thickness - 2);
-            }
-            gc.strokeLine(thickness + i * spacing, 0, thickness + i * spacing, lineHeight);
-        }
     }
 
     public static synchronized BufferedImage raster(final WebAppModel model) throws Throwable {
